requirements.txt
python-telegram-bot
feedparser==6.0.11
urllib3==1.26.16
